Add extra U_64 data field to Indexable Object Headers
Preliminary work to incorporate extra 64 bit field to
Indexable object headers. In case of contiguous arraylets the
extra field will hold the address of where the data starts,
which is right after the header. In case of discontiguous
arraylets the extra field will hold the address of the contiguous
double mapped region (assuming double map is enabled).
This does not apply to hybrid arraylets because they are non
existent when double mapping is enabled.

This is just for testing purposes. Eventually we’ll create new
Indexable Object header structures that contain this extra
64 bit field; at runtime we’ll decide which headers to use (e.g. if GC
policy is balanced and double mapping is enabled the runtime
decides to use the headers that contain the extra field).

Signed-off-by: Igor Braga <higorb1@gmail.com>