Disable double mapping when page size exceeds default page size
In systems that support huge pages, the heap is created using
the available huge pages (normally 2MB). The path taken to
create a heap with huge pages is different from the one taken
when normal page sizes are used. In order for double map to
work, heap must be created using mmap with a file descriptor
associated to the heap. However, early versions of glibc
only supports mmap + MAP_HUGETLB with anonymous mapping,
which is the opposite of what we want. For now we disable
double mapping whenever huge pages are used; however in a
later release we optimally want to enable it by using
memfd_create(2) which also returns a file descriptor just
like shm_open(3); however, memfd_create(2) has the same
semantics as other anonymous memory allocations

Signed-off-by: Igor Braga <higorb1@gmail.com>