Introduce concurrent copy forward functionality
This is phase 2 of #12089 . Here we introduce
MM_ConcurrentCopyForwardSchemeTask which is the equivalent
of Concurrent Scavenger MM_ConcurrentScavengeTask. We also
introduce CopyForward state machine that will dictate
concurrent copy forward phase transitions.

Draft: here we also try to simulate phase 3 of #12089, where
we finally make use of the newly introduced
MM_ConcurrentCopyForwardSchemeTask as well as the concurrent
copy forward state machine. Here we divide what was previously
one STW phase into 3 separate STW phases (STW -> STW -> STW).
There are a lot of hacks in place to make this possible, for
example assumeExclusiveVMAccess(), setWorkSTWDone() and others.
We also need to revisit _cycleState from MM_IncrementalGenerationalGC
and _persistentGlobalMarkPhaseState

Signed-off-by: Igor Braga <higorb1@gmail.com>