Prototype to make discontiguous double mapped arraylet look contiguous
Update effective address functions to always use dataAddr, which
point to the data associated with arraylet. And camouflage
discontiguous double mapped arraylets to look as if they were
InlineContiguous arraylets, with dataAddr pointing to double
mapped contiguous region. In that case arraylet leaves protection
are set to PROT_NONE, and reversed to PROT_WRITE and PROT_READ
when collected.

Signed-off-by: Igor Braga <higorb1@gmail.com>